Moved maths related files to Math folder inside utils Updated related unit tests
