+++ 100-js-algo/arithmeticExpression.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line max-params
 function arithmeticExpression(operand1, operand2, result) {
-    const { getInteger } = require("./utils/Math");
+    const { getInteger } = require("./utils/Math/Math");
     if (getInteger(operand1 + operand2) === result) {
         return true;
     } else if ((getInteger(operand1 - operand2) === result) || getInteger(operand2 - operand1) === result) {
