+++ 100-js-algo/utils/Math/Math.js
