+++ 100-js-algo/utils/Math/__test__/math.test.js
